Extract `utbot-spring-framework` module from `utbot-framework-module`
Remove duplicated `include("utbot-spring-framework")` in `settings.gradle.kts`
