+++ utbot-framework/src/main/kotlin/org/utbot/framework/codegen/domain/builtin/MockitoBuiltins.kt
@@ -20,21 +20,6 @@ internal val mockitoClassId = BuiltinClassId(
     simpleName = "Mockito",
 )
 
-internal val mockClassId = BuiltinClassId(
-    canonicalName = "org.mockito.Mock",
-    simpleName = "Mock",
-)
-
-internal val spyClassId = BuiltinClassId(
-    canonicalName = "org.mockito.Spy",
-    simpleName = "Spy"
-)
-
-internal val injectMocksClassId = BuiltinClassId(
-    canonicalName = "org.mockito.InjectMocks",
-    simpleName = "InjectMocks",
-)
-
 internal val ongoingStubbingClassId = BuiltinClassId(
     canonicalName = "org.mockito.stubbing.OngoingStubbing",
     simpleName = "OngoingStubbing",
